+++ _benchmarks/cma_solver/main.go
@@ -2,13 +2,13 @@ package main
 
 import (
 	"github.com/c-bata/goptuna"
-	"github.com/c-bata/goptuna/cma"
+	"github.com/c-bata/goptuna/cmaes"
 	kurobako "github.com/sile/kurobako-go"
 	"github.com/sile/kurobako-go/goptuna/solver"
 )
 
 func createStudy(seed int64) (*goptuna.Study, error) {
-	relativeSampler := cma.NewSampler(cma.SamplerOptionSeed(seed))
+	relativeSampler := cmaes.NewSampler(cmaes.SamplerOptionSeed(seed))
 	return goptuna.CreateStudy("example-study",
 		goptuna.StudyOptionRelativeSampler(relativeSampler))
 }
